fajita: overlay: Increase rounded mask size to actually fit corners
After increasing this value, the Google Assistant gesture animation now
glides along the bottom corners of the screen, as opposed to being
almost entirely out of bounds from what the user can see. It is still
not a 100% perfect fit, but that would probably require a way to define
a custom svg path for the animation to follow.

Change-Id: Ib99c1f70944cc1673cee379620fd3cf939b4fa72